refactor: add subtest command
